+++ Frontend/src/lib/components/homepage/Capabilities.svelte
@@ -1,0 +1,24 @@
+<div
+  class="animate-border w-full max-w-[422px] rounded-2xl border border-transparent [background:linear-gradient(45deg,#172033,theme(colors.slate.800)_50%,#172033)_padding-box,conic-gradient(from_var(--border-angle),theme(colors.slate.600/.48)_80%,_theme(colors.indigo.500)_86%,_theme(colors.indigo.300)_90%,_theme(colors.indigo.500)_94%,_theme(colors.slate.600/.48))_border-box]">
+  <div class="p-5">
+    <div class="flex items-center justify-between">
+      <div>
+        <div class="mb-0.5 flex items-center space-x-2">
+          <div class="mb-1 text-2xl font-bold text-slate-200">17,479</div>
+          <div class="text-xs font-medium text-emerald-500">+48%</div>
+        </div>
+        <div class="text-sm font-medium text-slate-500">Monthly visits</div>
+      </div>
+      <button class="flex h-8 w-8 items-center justify-center text-slate-500 hover:text-slate-400">
+        <span class="sr-only">Open menu</span>
+        <svg class="fill-current" xmlns="http://www.w3.org/2000/svg" width="16" height="4" fill="none">
+          <path
+            d="M8 4a2 2 0 1 0 0-4 2 2 0 0 0 0 4ZM2 4a2 2 0 1 0 0-4 2 2 0 0 0 0 4ZM14 4a2 2 0 1 0 0-4 2 2 0 0 0 0 4Z" />
+        </svg>
+      </button>
+    </div>
+  </div>
+  <div class="px-5">
+    <img src="https://images.unsplash.com/photo-1500917293891-ef795e70e1f6" width="380" height="173" alt="Card" />
+  </div>
+</div>