+++ Frontend/src/routes/+page.svelte
@@ -1,9 +1,8 @@
 <script lang="ts">
-  import {BackgroundBeams, CheveronImg, CheckmarkImg, Slogan} from '$components';
+  import {BackgroundBeams, CheveronImg, Slogan, Card3D, PricingTable, FeatureGrid, Capabilities} from '$components';
   import {notify, changePricingModel} from '$lib';
   import type {Notification} from '$types';
   import {satisfaction} from '$images';
-  import {pricingModel} from '$store';
   import {enhance} from '$app/forms';
 
   export let form: Notification;
@@ -17,7 +16,7 @@
     content="ShadowSelf is a platform that safeguard you and your sensitive data by creating sythetic identities that can be used to register and authenticate while concealing your actual identity from being at risk of misuse, breach, theft, or fraud." />
 </svelte:head>
 
-<section id="catch">
+<section id="slogan">
   <Slogan />
   <p class="my-6 w-1/2 text-balance text-center">
     Experience the freedom of online interactions without the fear of compromise. Our platform provides the tools to
@@ -27,13 +26,15 @@
   <div class="flex gap-8">
     <a href="#product" class="no-underline">
       <button class="alt group flex items-center gap-1">
-        Learn More<CheveronImg className="group-hover:rotate-90" /></button>
+        Learn More<CheveronImg className="group-hover:rotate-90" />
+      </button>
     </a>
     <a href="/signup"><button>Get Started!</button></a>
   </div>
 </section>
 
 <section id="product">
+  <Card3D />
   <div class="flex flex-col items-start gap-8">
     <h1 class="text-6xl">Reclaim Control of Your Data.</h1>
     <p>
@@ -45,9 +46,9 @@
   </div>
 </section>
 
-<section id="features"><h1>Features</h1></section>
+<section id="capabilities"><Capabilities /></section>
 
-<section id="comparaisons"><h1>Comparaisons</h1></section>
+<section id="features"><FeatureGrid /></section>
 
 <section id="pricing">
   <div id="pricing-model">
@@ -55,39 +56,7 @@
     <button id="annually" on:click={() => changePricingModel('annually')}>Annually</button>
     <button id="lifetime" on:click={() => changePricingModel('lifetime')}>Lifetime</button>
   </div>
-  <div id="pricing-card">
-    <h3>{$pricingModel.title}</h3>
-    <div class="flex flex-col items-center gap-4 px-16 py-6 text-center">
-      <p class="-mb-7 text-neutral-600 line-through">${$pricingModel.price}</p>
-      <h1 class="flex items-start text-6xl">
-        <span class="mt-3 text-4xl">$</span>
-        {$pricingModel.afterDiscount}
-      </h1>
-      <p class="-mt-6 mb-8 text-neutral-400">{$pricingModel.description}</p>
-      <p>What will be included :</p>
-      <div class="flex gap-8 py-4 text-left text-xl leading-10">
-        <ul class="list-inside">
-          <li><CheckmarkImg />Custom Identity</li>
-          <li><CheckmarkImg />Email Address</li>
-          <li><CheckmarkImg />Phone Number</li>
-        </ul>
-        <ul class="mx-8 list-inside">
-          <li><CheckmarkImg />Virtual Card</li>
-          <li><CheckmarkImg />Crypto Wallet</li>
-          <li><CheckmarkImg />VPN Access</li>
-        </ul>
-        <ul class="list-inside">
-          <li><CheckmarkImg />Account Management</li>
-          <li><CheckmarkImg />Money-Back Guarantee</li>
-          <li><CheckmarkImg />24/7 Support</li>
-        </ul>
-      </div>
-      <a href="/signup">
-        <button id="purchase">Purchase<CheveronImg /></button>
-      </a>
-      <p class="text-neutral-500">All major crypto currencies are supported.</p>
-    </div>
-  </div>
+  <PricingTable />
 </section>
 
 <section id="reflection">
@@ -100,7 +69,7 @@
       reliable, secure, and incredibly convenient to use.
     </p>
   </div>
-  <img class="animate-shake-up w-full min-w-[30vw]" src={satisfaction} alt="Customer satisfaction" />
+  <img class="animate-shake w-full min-w-[30vw]" src={satisfaction} alt="Customer satisfaction" />
 </section>
 
 <section id="waitlist">
@@ -144,27 +113,8 @@
     @apply text-primary-700 from-neutral-300 to-[#b3bdcc];
   }
 
-  #pricing-card {
-    @apply rounded-2xl border-4 border-neutral-400 shadow-2xl shadow-neutral-900;
-    @apply bg-gradient-to-br from-[#0e1526] to-[#0a0f1b];
-    @apply hover:shadow-8xl transition-all duration-500 ease-in-out hover:-translate-y-4;
-  }
-
-  #purchase {
-    @apply my-4 flex w-full items-center justify-center gap-2 p-6 text-3xl font-bold;
-  }
-
   #waitlist p {
     @apply z-10 w-1/3 hyphens-none text-balance text-center text-lg text-neutral-400;
-  }
-
-  h3 {
-    @apply from-primary-600 to-primary-700 bg-gradient-to-b shadow-md shadow-neutral-950;
-    @apply w-full rounded-none rounded-t-xl border-b border-neutral-400 p-8 text-center text-3xl font-bold;
-  }
-
-  li {
-    @apply flex items-center gap-4;
   }
 
   input {